LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/redirect/redirect-http-upgrade.sub-expected.txt

Upgraded font


Harness Error (TIMEOUT), message = null

PASS Http upgrade iframe
PASS Http upgrade top level navigation
TIMEOUT Http upgrade embed Test timed out
NOTRUN Http upgrade fetch() api
NOTRUN Http upgrade object
PASS Http upgrade preload
NOTRUN Http upgrade stylesheet
NOTRUN Http upgrade track
NOTRUN Http upgrade image => No headers
PASS Http upgrade script => No headers
FAIL Http upgrade script => No headers: sec-fetch-dest assert_equals: expected "script" but got ""
FAIL Http upgrade script => No headers: sec-fetch-mode assert_equals: expected "no-cors" but got ""
FAIL Http upgrade script => No headers: sec-fetch-site assert_equals: expected "cross-site" but got ""
PASS Http upgrade script => No headers: sec-fetch-user
FAIL Http upgrade iframe: sec-fetch-dest assert_equals: expected "iframe" but got ""
FAIL Http upgrade iframe: sec-fetch-mode assert_equals: expected "navigate" but got ""
FAIL Http upgrade iframe: sec-fetch-site assert_equals: expected "cross-site" but got ""
PASS Http upgrade iframe: sec-fetch-user
NOTRUN Http upgrade font => No headers
FAIL Http upgrade top level navigation: sec-fetch-dest assert_equals: expected "document" but got ""
FAIL Http upgrade top level navigation: sec-fetch-mode assert_equals: expected "navigate" but got ""
FAIL Http upgrade top level navigation: sec-fetch-site assert_equals: expected "cross-site" but got ""
PASS Http upgrade top level navigation: sec-fetch-user
